feature(create-mochData-file):add mockData js file
[Finishes #168898013]
