+++ test/mockData/mockData.js
@@ -1,0 +1,14 @@
+/* eslint-disable linebreak-style */
+const payload = {
+  id: 2,
+  first_name: 'Patrick',
+  last_name: 'Tunezerwane',
+  email: 'tp3@gmail.com',
+  password: '$2b$10$Atnw/KEDHvmcSdNTRWfMfOZIOOQFOIynwjiYqGGZx3xtemaF6NGe6',
+  gender: 'Male',
+  jobRole: 'Manager',
+  department: 'ICT',
+  address: 'Kigali',
+  is_admin: false,
+};
+module.exports = payload;
